Fixed files.size type to long. int is too small

+++ repo_models/files.py
@@ -177,7 +177,7 @@ FIELDS = [
     
     {
         'name':       'size',
-        'model_type': int,
+        'model_type': long,
         'default':    None,
         'csv': {
             'export': 'ignore',
